Clean checker component rendering
