+++ src/renderer/pages/Home.tsx
@@ -2,12 +2,12 @@ import { useEffect, useState } from 'react';
 import { Space } from 'antd';
 import CHANNELS from '../utils/channels';
 import eventHandler from '../utils/eventHandler';
-import { FileKeyType, CheckerTypeType } from '../utils/Types';
+import { FileKeyType } from '../utils/Types';
 import FileSelectionRow from '../components/FileSelectionRow';
 import NumberSelectionRow from '../components/NumberSelectionRow';
 import JudgeButton from '../components/JudgeButton';
 import Results from '../components/Results';
-import DropdownSelectorRow from '../components/DropdownSelectorRow';
+import CheckerTypeSelectorRow from '../components/CheckerTypeSelectorRow';
 
 type FileData = {
   [K in FileKeyType]?: string;
@@ -20,17 +20,8 @@ const FILE_KEYS = {
   OUTPUT: 'output',
 };
 
-const CHECKER_TYPE_OPTIONS: Array<CheckerTypeType> = [
-  'diff',
-  'token',
-  'epsilon',
-];
-
 export default function Home() {
   const [fileInfo, setFileInfo] = useState<FileData>({});
-  const [checkerType, setCheckerType] = useState<CheckerTypeType>(
-    CHECKER_TYPE_OPTIONS[0]
-  );
 
   // This is ok because the setter is only called as a LISTENER
   // eslint-disable-next-line react-hooks/exhaustive-deps
@@ -52,15 +43,6 @@ export default function Home() {
 
   const onChangeTimeLimit = (limit: number) => {
     eventHandler.setTimeLimit(limit);
-  };
-
-  const onChangeCheckerType = (newCheckerType: CheckerTypeType) => {
-    setCheckerType(newCheckerType);
-    eventHandler.setCheckerType(newCheckerType);
-  };
-
-  const onChangeEpsilon = (newEpsilon: number) => {
-    eventHandler.setEpsilon(newEpsilon);
   };
 
   return (
@@ -92,22 +74,7 @@ export default function Home() {
           units="seconds"
           onChange={onChangeTimeLimit}
         />
-        <DropdownSelectorRow
-          label="Checker"
-          choices={CHECKER_TYPE_OPTIONS}
-          onSelect={(value: string) =>
-            onChangeCheckerType(value as CheckerTypeType)
-          }
-        />
-        {checkerType === 'epsilon' ? (
-          <NumberSelectionRow
-            label="Epsilon"
-            type="float"
-            min={0}
-            defaultValue={0.000001}
-            onChange={onChangeEpsilon}
-          />
-        ) : null}
+        <CheckerTypeSelectorRow />
         <JudgeButton />
         <Results />
       </Space>
